requirements.txt
aiogram
paddleocr
pillow
sqlalchemy
asyncpg
psycopg2
